+++ test/rounding.jl
@@ -63,13 +63,13 @@ end
                   pi,ℯ,eulergamma,catalan,golden,
                   typemax(Int64),typemax(UInt64),typemax(Int128),typemax(UInt128),0xa2f30f6001bb2ec6]
             pn = T(v,RoundNearest)
-            @test pn == convert(T,BigFloat(v))
+            @test pn == convert(T,Float64(v))
             pz = T(v,RoundToZero)
-            @test pz == setrounding(()->convert(T,BigFloat(v)), BigFloat, RoundToZero)
+            @test pz == setrounding(()->convert(T,Float64(v)), Float64, RoundToZero)
             pd = T(v,RoundDown)
-            @test pd == setrounding(()->convert(T,BigFloat(v)), BigFloat, RoundDown)
+            @test pd == setrounding(()->convert(T,Float64(v)), Float64, RoundDown)
             pu = T(v,RoundUp)
-            @test pu == setrounding(()->convert(T,BigFloat(v)), BigFloat, RoundUp)
+            @test pu == setrounding(()->convert(T,Float64(v)), Float64, RoundUp)
 
             @test pn == pd || pn == pu
             @test v > 0 ? pz == pd : pz == pu
